various errors on MS from QGLviewer activation

+++ src/DGtal/io/viewers/Viewer3D.h
@@ -1331,7 +1331,7 @@ namespace DGtal
     float myGLScaleFactorZ;
 
     // Used to apply interactive light rotation
-    float myLigthRotationStep; /// the angle rotation increment used for interactive light move
+    double myLigthRotationStep; /// the angle rotation increment used for interactive light move
     int myRefMouseXPos; /// the reference mouse x-position used to determince the light position change (azimuth)
     int myRefMouseYPos; /// the reference mouse y-position used to determince the light position change (inclination)
     bool myIsMovingLight; /// flag to display the ligth source when it is moved by the user
